Equivilance Classes Introduced
As far as anagrams are concerned, two words with the same letters may aswell be the exact same.

To deal with this, I've slotted every word into an equivilance class of other words with the same number of letters, and done all of the main anagram computing on these classes.

This should, for the same amount of computation, give us approximately G^D times as many solutions, where G is the geometric mean of the number of elements in an equivilance class, and D is the depth of an anagram search.